+++ lib/src/dart_randomx_base.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:ffi' as ffi;
-import 'dart:io' show Platform, Directory, File;
+import 'dart:io' show Directory, File, Platform, Process;
 import 'dart:isolate';
 import 'dart:typed_data';
 
@@ -25,9 +25,17 @@ typedef FunctionRandomxHash = void Function(int id,
 typedef FFIFunctionRandomxDestroy = ffi.Void Function(ffi.Int32 id);
 typedef FunctionRandomxDestroy = void Function(int id);
 
+/// Dart `RandomX` library wrapper.
 class RandomX {
+  /// Returns `true` if the library is already loaded.
   static bool get isLibLoaded => _dynLib != null;
 
+  /// Returns the path of the loaded library.
+  ///
+  /// See [isLibLoaded] as [loadLib].
+  static String? get loadedLibraryPath => _dynLibPath;
+
+  /// Loads the wrapper library.
   static Future<bool> loadLib() async {
     if (isLibLoaded) return true;
     return await _loadLib();
@@ -35,6 +43,7 @@ class RandomX {
 
   static Completer<bool>? _loadLibCompleter;
   static ffi.DynamicLibrary? _dynLib;
+  static String? _dynLibPath;
 
   static Future<bool> _loadLib() async {
     var completer = _loadLibCompleter;
@@ -49,11 +58,30 @@ class RandomX {
     }
 
     try {
-      var libraryPath = await _getLibraryPath();
-      var dynLib = ffi.DynamicLibrary.open(libraryPath);
+      var libraryPaths = await _getLibraryPath();
+      ffi.DynamicLibrary? dynLib;
+      String? dynLibPath;
+
+      Object? error;
+      for (var libPath in libraryPaths) {
+        try {
+          dynLib = ffi.DynamicLibrary.open(libPath);
+          dynLibPath = libPath;
+          break;
+        } catch (e) {
+          print('$libPath >> $e');
+          error = e;
+        }
+      }
+
+      if (dynLib == null) {
+        throw StateError("Can't find library at $libraryPaths > $error");
+      }
+
       _mapLibFunctions(dynLib);
 
       _dynLib = dynLib;
+      _dynLibPath = dynLibPath;
       completer.complete(true);
     } catch (e, s) {
       completer.completeError(e, s);
@@ -85,11 +113,13 @@ class RandomX {
   static final String _libraryHeaderFile = 'wrapper_randomx.h';
   static final String _libraryName = 'libwrapper_randomx';
 
-  static Future<String> _getLibraryPath() async {
-    var libExt = _getPlatformExtension();
+  static Future<List<String>> _getLibraryPath() async {
+    var libExtensions = _getPlatformExtensions();
     var wrapperLibDir = await _findWrapperLibraryDirectory();
-    var libraryPath = path.join(wrapperLibDir.path, '$_libraryName$libExt');
-    return libraryPath;
+    var libraryPaths = libExtensions
+        .map((ext) => path.join(wrapperLibDir.path, '$_libraryName$ext'))
+        .toList();
+    return libraryPaths;
   }
 
   static Future<Directory> _findWrapperLibraryDirectory() async {
@@ -136,11 +166,32 @@ class RandomX {
     return null;
   }
 
-  static String _getPlatformExtension() {
-    if (Platform.isLinux) return '.so';
-    if (Platform.isMacOS) return '.dylib';
-    if (Platform.isWindows) return '.dll';
+  static List<String> _getPlatformExtensions() {
+    if (Platform.isLinux) {
+      return ['.so'];
+    } else if (Platform.isMacOS) {
+      var arm64 = isMacOSArm64();
+      var main = arm64 ? '-arm64.dylib' : '-x64.dylib';
+      return [main, '.dylib'];
+    } else if (Platform.isWindows) {
+      return ['.dll'];
+    }
+
     throw 'Unsupported platform ${Platform.operatingSystem}';
+  }
+
+  static bool isMacOSArm64() {
+    if (!Platform.isMacOS) return false;
+    try {
+      var process = Process.runSync('/usr/bin/uname', ['-a']);
+      var output = '${process.stdout}'.toLowerCase();
+      var idx = output.lastIndexOf('kernel');
+      var rest = idx >= 0 ? output.substring(idx) : output;
+      return rest.contains('arm64');
+    } catch (e) {
+      print(e);
+      return false;
+    }
   }
 
   static int _idCount = 0;
@@ -148,6 +199,9 @@ class RandomX {
   final int id = ++_idCount;
   late final int sizeOfHash;
 
+  /// [RandomX] default constructor.
+  ///
+  /// Call [RandomX.loadLib] before instantiate.
   RandomX() {
     if (!isLibLoaded) {
       throw StateError(
@@ -161,6 +215,7 @@ class RandomX {
 
   Uint8List? get initKey => _initKey;
 
+  /// Returns `true` if [key] equals to [initKey].
   bool equalsInitKey(Uint8List key) {
     var initKey = _initKey;
     if (initKey == null) {
@@ -172,8 +227,12 @@ class RandomX {
 
   bool? _fullMemory;
 
+  /// Returns `true` if [init] was called with `fullMemory`.
   bool? get fullMemory => _fullMemory;
 
+  /// Initializes the `RandomX` instance with [key].
+  ///
+  /// - [fullMemory] if `true` will use 2G+ of memory (fast mode).
   void init(Uint8List key, {bool fullMemory = false}) {
     _initKey = UnmodifiableUint8ListView(Uint8List.fromList(key));
     var keyPointer = key.asPointer();
@@ -212,6 +271,7 @@ class RandomX {
   ffi.Pointer<ffi.Uint8> _getOutputPointer() =>
       _hashOutputPointer ??= createBytesPointer(sizeOfHash);
 
+  /// Hash [bytes] and puts the result into [output].
   void hashTo(Uint8List bytes, Uint8List output) {
     var inputPointer = _getInputPointer(bytes.length);
     inputPointer.setBytes(bytes);
@@ -224,12 +284,14 @@ class RandomX {
     output.setAll(0, outBytes);
   }
 
+  /// Hash [bytes] and returns the result as a [Uint8List].
   Uint8List hash(Uint8List bytes) {
     var output = createHashBytesArray();
     hashTo(bytes, output);
     return output;
   }
 
+  /// Destroys `this` `RandomX` instance resources.
   void destroy() => _functionRandomxDestroy(id);
 
   @override